part2/02-streamlit/requirements.txt
albumentations==1.1.0
efficientnet-pytorch==0.7.1
matplotlib==3.2.1
opencv-python==4.5.1.48
streamlit==1.1.0
torch==1.10.0
torchvision==0.11.1
PyYAML==6.0